+++ apps/docs/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @pie-qti/app-docs
 
+## 0.1.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @pie-qti/assessment-player@0.1.4
+  - @pie-qti/item-player@0.1.4
+
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @pie-qti/assessment-player@0.1.3
+  - @pie-qti/item-player@0.1.3
+
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @pie-qti/assessment-player@0.1.2
+  - @pie-qti/item-player@0.1.2
+
 ## 0.1.1
 
 ### Patch Changes