+++ apps/transform/CHANGELOG.md
@@ -1,5 +1,62 @@
 # @pie-qti/app-transform
 
+## 0.1.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @pie-qti/assessment-player@0.1.4
+  - @pie-qti/default-components@0.1.4
+  - @pie-qti/demo-vendor-extensions@0.1.4
+  - @pie-qti/i18n@0.1.4
+  - @pie-qti/item-player@0.1.4
+  - @pie-qti/player-elements@0.1.4
+  - @pie-qti/qti-common@0.1.4
+  - @pie-qti/storage@0.1.4
+  - @pie-qti/to-pie@0.1.4
+  - @pie-qti/transform-core@0.1.4
+  - @pie-qti/transform-types@0.1.4
+  - @pie-qti/typeset-katex@0.1.4
+  - @pie-qti/web-component-loaders@0.1.4
+
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @pie-qti/assessment-player@0.1.3
+  - @pie-qti/default-components@0.1.3
+  - @pie-qti/demo-vendor-extensions@0.1.3
+  - @pie-qti/i18n@0.1.3
+  - @pie-qti/item-player@0.1.3
+  - @pie-qti/player-elements@0.1.3
+  - @pie-qti/qti-common@0.1.3
+  - @pie-qti/storage@0.1.3
+  - @pie-qti/to-pie@0.1.3
+  - @pie-qti/transform-core@0.1.3
+  - @pie-qti/transform-types@0.1.3
+  - @pie-qti/typeset-katex@0.1.3
+  - @pie-qti/web-component-loaders@0.1.3
+
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @pie-qti/assessment-player@0.1.2
+  - @pie-qti/default-components@0.1.2
+  - @pie-qti/demo-vendor-extensions@0.1.2
+  - @pie-qti/i18n@0.1.2
+  - @pie-qti/item-player@0.1.2
+  - @pie-qti/player-elements@0.1.2
+  - @pie-qti/qti-common@0.1.2
+  - @pie-qti/storage@0.1.2
+  - @pie-qti/to-pie@0.1.2
+  - @pie-qti/transform-core@0.1.2
+  - @pie-qti/transform-types@0.1.2
+  - @pie-qti/typeset-katex@0.1.2
+  - @pie-qti/web-component-loaders@0.1.2
+
 ## 0.1.1
 
 ### Patch Changes