+++ packages/core/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @pie-qti/transform-core
 
+## 0.1.4
+
+### Patch Changes
+
+- Temporary release changeset: patch all publishable packages to keep lockstep versions.
+- Updated dependencies
+  - @pie-qti/element-schemas@0.1.4
+  - @pie-qti/logger@0.1.4
+  - @pie-qti/transform-types@0.1.4
+
+## 0.1.3
+
+### Patch Changes
+
+- Temporary release changeset: patch all publishable packages to keep lockstep versions.
+- Updated dependencies
+  - @pie-qti/element-schemas@0.1.3
+  - @pie-qti/logger@0.1.3
+  - @pie-qti/transform-types@0.1.3
+
+## 0.1.2
+
+### Patch Changes
+
+- Temporary release changeset: patch all publishable packages to keep lockstep versions.
+- Updated dependencies
+  - @pie-qti/element-schemas@0.1.2
+  - @pie-qti/logger@0.1.2
+  - @pie-qti/transform-types@0.1.2
+
 ## 0.1.1
 
 ### Patch Changes