+++ packages/player-elements/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @pie-qti/player-elements
 
+## 0.1.4
+
+### Patch Changes
+
+- Temporary release changeset: patch all publishable packages to keep lockstep versions.
+- Updated dependencies
+  - @pie-qti/assessment-player@0.1.4
+  - @pie-qti/item-player@0.1.4
+
+## 0.1.3
+
+### Patch Changes
+
+- Temporary release changeset: patch all publishable packages to keep lockstep versions.
+- Updated dependencies
+  - @pie-qti/assessment-player@0.1.3
+  - @pie-qti/item-player@0.1.3
+
+## 0.1.2
+
+### Patch Changes
+
+- Temporary release changeset: patch all publishable packages to keep lockstep versions.
+- Updated dependencies
+  - @pie-qti/assessment-player@0.1.2
+  - @pie-qti/item-player@0.1.2
+
 ## 0.1.1
 
 ### Patch Changes