+++ packages/to-pie/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @pie-qti/to-pie
 
+## 0.1.4
+
+### Patch Changes
+
+- Temporary release changeset: patch all publishable packages to keep lockstep versions.
+- Updated dependencies
+  - @pie-qti/ims-cp-core@0.1.4
+  - @pie-qti/ims-cp-node@0.1.4
+  - @pie-qti/transform-core@0.1.4
+  - @pie-qti/transform-types@0.1.4
+
+## 0.1.3
+
+### Patch Changes
+
+- Temporary release changeset: patch all publishable packages to keep lockstep versions.
+- Updated dependencies
+  - @pie-qti/ims-cp-core@0.1.3
+  - @pie-qti/ims-cp-node@0.1.3
+  - @pie-qti/transform-core@0.1.3
+  - @pie-qti/transform-types@0.1.3
+
+## 0.1.2
+
+### Patch Changes
+
+- Temporary release changeset: patch all publishable packages to keep lockstep versions.
+- Updated dependencies
+  - @pie-qti/ims-cp-core@0.1.2
+  - @pie-qti/ims-cp-node@0.1.2
+  - @pie-qti/transform-core@0.1.2
+  - @pie-qti/transform-types@0.1.2
+
 ## 0.1.1
 
 ### Patch Changes